disable password
Evidently Github pages disabled the ability to run javascript, so the password protection feature no longer worked. This was also true for the demo upon which I based the password protection. All this update accomplishes is moving all site data outside of the password hash directory to the root of the site.